tests: factor out partition metrics helper
